draft with new updates and validation checks finished, now to test

+++ validator_cleanup.rb
@@ -2,13 +2,13 @@ require 'fileutils'
 require 'logger'
 require 'find'
 
-# ---------------------- VARIABLES
+# ---------------------- VARIABLES (HEADER)
 unescapeargv = ARGV[0].chomp('"').reverse.chomp('"').reverse
 input_file = File.expand_path(unescapeargv)
 input_file = input_file.split(Regexp.union(*[File::SEPARATOR, File::ALT_SEPARATOR].compact)).join(File::SEPARATOR)
 filename_split = input_file.split(Regexp.union(*[File::SEPARATOR, File::ALT_SEPARATOR].compact)).pop
 input_file_normalized = input_file.gsub(/ /, "")
-filename_normalized = filename_split.gsub(/ /, "")
+filename_normalized = filename_split.scan( /[^[:alnum:]\._-]/ ) { |badchar| filename_split=filename_split.tr(badchar,'') } #strip any chars but alphanumeric, plus these 3 .-_
 basename_normalized = File.basename(filename_normalized, ".*")
 extension = File.extname(filename_normalized)
 project_dir = input_file.split(Regexp.union(*[File::SEPARATOR, File::ALT_SEPARATOR].compact))[0...-2].join(File::SEPARATOR)
@@ -16,52 +16,65 @@ project_name = input_file.split(Regexp.union(*[File::SEPARATOR, File::ALT_SEPARA
 inbox = File.join(project_dir, 'IN')
 outbox = File.join(project_dir, 'OUT')
 working_dir = File.join('S:', 'validator_tmp')
-tmp_dir=File.join(working_dir, basename_normalized) 
+tmp_dir=File.join(working_dir, basename_normalized)
 working_file = File.join(tmp_dir, filename_normalized)
-done_file = File.join(tmp_dir, "#{basename_normalized}_DONE#{extension}")
+bookinfo_file = File.join(tmp_dir,'book_info.json')
+stylecheck_file = File.join(tmp_dir,'style_check.json')
+submitter_file = File.join(tmp_dir,'submitter.json')
+testing_value_file = File.join("C:", "staging.txt")
 inprogress_file = File.join(inbox,"#{filename_normalized}_IN_PROGRESS.txt")
-errlog = false
-errfile = ''
-err_notice = File.join(outbox,"ERROR--#{filename_normalized}--Validator_Failed.txt")
-
+errFile = File.join(inbox, "ERROR_RUNNING_#{filename_normalized}.txt")
 
 # ---------------------- LOGGING
 logfolder = File.join(working_dir, 'logs')
-logfile = File.join(logfolder, "#{basename_normalized}_log.txt") 
+logfile = File.join(logfolder, "#{basename_normalized}_log.txt")
 logger = Logger.new(logfile)
 logger.formatter = proc do |severity, datetime, progname, msg|
   "#{datetime}: #{progname} -- #{msg}\n"
 end
 
+# ---------------------- LOCAL VARIABLES
+err_notice = File.join(outbox,"ERROR--#{filename_normalized}--Validator_Failed.txt")
+done_file = File.join(tmp_dir, "#{basename_normalized}_DONE#{extension}")
+errlog = false
+timestamp = Time.now.strftime('%Y-%m-%d_%H-%M-%S')
+backup_tmp_dir = File.join(logfolder, "#{tmp_dir}_#{timestamp}")
+
 
 #--------------------- RUN
 #check for errlog in tmp_dir:
 Find.find(tmp_dir) { |file|
-	if file =~ /^.*\.(txt|json|log)/ && file !~ /^.*userinfo.json/
-		logger.info('validator_mailer') {"error log found in tmpdir: #{file}, setting email text accordingly"}
+	if file != stylecheck_file && file != bookinfo_file && file != working_file && file != submitter_file
+		logger.info('validator_cleanup') {"error log found in tmpdir: #{file}"}
 		errlog = true
-		errfile = file
 	end
 }
 
 if errlog
-	logger.info('validator_cleanup') {"a major error was detected while running macros on \"#{filename_normalized}\""}
-	FileUtils.cp(errfile, logfolder)   #copy error logfile to logdir
-	FileUtils.rm_rf tmp_dir     #optionally still delete tmp_dir for file:
-	#TODO here:  invoke mailer / alert for us. or have that as separate script.
-	# or maybe that's a separate .rb script that runs prior to this one if errfile exists
-	FileUtils.mv input_file, outbox	  #returm the original file
+	logger.info('validator_cleanup') {"a major error(ALERT) was detected while running macros on \"#{filename_normalized}\", moving tmpdir to logfolder for further study"}
+	FileUtils.mv tmp_dir, backup_tmp_dir  
+	FileUtils.mv input_file, outbox	  #returm the original file to user
     File.open(err_notice, 'w') { |f|
         f.puts "Bookmaker_validator failed for file #{filename_normalized}. Please email workflows@macmillan.com for help!"
     }
-else
-	if Dir.exist?(tmp_dir)
-		logger.info('validator_cleanup') {"file \"#{filename_normalized}\" processing complete, renaming and moving to OUT folder"}
-		File.rename(working_file, done_file)
-		FileUtils.cp_r "#{tmp_dir}/.", outbox
-		FileUtils.mv input_file, outbox
-		FileUtils.rm_rf tmp_dir
-		logger.info('validator_cleanup') {"processing of \"#{filename_normalized}\" completed"}
-	end
+    logger.info('validator_cleanup') {"returned orig file and error notice to user, exitng cleanup"}
+elsif Dir.exist?(tmp_dir)
+	if !File.file?(bookinfo_file)
+		FileUtils.mv input_file, outbox	  #returm the original file to user
+    	File.open(err_notice, 'w') { |f|
+	        f.puts "isbn lookup failed for file #{filename_normalized}, bookmaker_validator could not run! Please double-check the isbn in your filename!"
+	    }
+	    logger.info('validator_cleanup') {"book_info.json missing, returned orig file and isbn lookup error notice to user, exitng cleanup"}
+	end	
+	logger.info('validator_cleanup') {"file \"#{filename_normalized}\" processing complete, renaming and moving to OUT folder"}
+	File.rename(working_file, done_file)
+	FileUtils.cp_r "#{tmp_dir}/.", outbox
+	FileUtils.mv input_file, outbox
+	FileUtils.rm_rf tmp_dir
+	logger.info('validator_cleanup') {"processing of \"#{filename_normalized}\" completed"}
+elsif File.file?(errFile)
+	logger.info('validator_cleanup') {"errFile found, indicating file failed basic validation, moving orig & errNotice from IN to OUT folder"}
+	FileUtils.mv input_file, outbox
+	FileUtils.mv errFile, outbox
 end
-FileUtils.rm inprogress_file
+if File.file?(inprogress_file) then FileUtils.rm inprogress_file end
